Event Router 增加顺序能力

+++ domain-support/src/main/java/com/stellariver/milky/domain/support/event/FinalEventRouter.java
@@ -17,6 +17,6 @@ public @interface FinalEventRouter {
 
     String executor() default "asyncExecutor";
 
-    double order() default Double.MAX_VALUE;
+    long order() default Long.MAX_VALUE;
 
 }
